add integration package for integration tests Signed-off-by: dbulahov <bulahovdenis@gmail.com>
