+++ itest/src/test/java/org/eclipse/daanse/jakarta/xml/ws/itests/integration/SoapWhiteboardTest.java
@@ -12,7 +12,7 @@
 *   Stefan Bischof (bipolis.org) - initial
 *   Christoph Läubrich - initial
 */
-package org.eclipse.daanse.jakarta.xml.ws.itests;
+package org.eclipse.daanse.jakarta.xml.ws.itests.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.fail;
