+++ itest/src/test/java/org/eclipse/daanse/jakarta/xml/ws/itests/integration/TestSoapHandler.java
@@ -12,34 +12,42 @@
 *   Stefan Bischof (bipolis.org) - initial
 *   Christoph Läubrich - initial
 */
-package org.eclipse.daanse.jakarta.xml.ws.itests;
+package org.eclipse.daanse.jakarta.xml.ws.itests.integration;
 
+import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import jakarta.xml.ws.handler.LogicalHandler;
-import jakarta.xml.ws.handler.LogicalMessageContext;
-import jakarta.xml.ws.handler.MessageContext;
+import javax.xml.namespace.QName;
 
-public class TestLogicalHandler implements LogicalHandler<LogicalMessageContext> {
+import jakarta.xml.ws.handler.MessageContext;
+import jakarta.xml.ws.handler.soap.SOAPHandler;
+import jakarta.xml.ws.handler.soap.SOAPMessageContext;
+
+public class TestSoapHandler implements SOAPHandler<SOAPMessageContext> {
 
     AtomicInteger handledMessages = new AtomicInteger();
 
     @Override
-    public boolean handleMessage(LogicalMessageContext context) {
+    public boolean handleMessage(SOAPMessageContext context) {
         int msg = handledMessages.incrementAndGet();
-        System.out.println("TestLogicalHandler.handleMessage no. " + msg);
+        System.out.println("TestSoapHandler.handleMessage no. " + msg);
         return true;
     }
 
     @Override
-    public boolean handleFault(LogicalMessageContext context) {
-        System.out.println("TestLogicalHandler.handleFault()");
+    public boolean handleFault(SOAPMessageContext context) {
+        System.out.println("TestSoapHandler.handleFault()");
         return true;
     }
 
     @Override
     public void close(MessageContext context) {
-        System.out.println("TestLogicalHandler.close()");
+        System.out.println("TestSoapHandler.close()");
+    }
+
+    @Override
+    public Set<QName> getHeaders() {
+        return Set.of();
     }
 
 }
